refactor all nav btns to common navBtn component, moved editPattern to containers
